feat(wip): job proposal plugins structure
