+++ packages/plugins/job-proposal/.dockerignore
