+++ packages/plugins/job-proposal/README.md
@@ -1,0 +1,15 @@
+# Job Proposal Plugin
+
+## Overview
+
+## Features
+
+## Installation
+
+To install the Job Proposal Plugin, simply run the following command in your terminal:
+
+```bash
+npm install @gauzy/job-proposal-plugin
+# or
+yarn add @gauzy/job-proposal-plugin
+```
