+++ packages/plugins/job-proposal/jest.config.js
@@ -1,9 +1,9 @@
 module.exports = {
-	name: 'job-proposals',
+	name: 'job-proposal',
 	preset: '../../../jest.config.js',
 	transform: {
 		'^.+\\.[tj]sx?$': 'ts-jest'
 	},
 	moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
-	coverageDirectory: '../../coverage/plugins/job-proposals'
+	coverageDirectory: '../../coverage/plugins/job-proposal'
 };
