+++ packages/plugins/job-proposal/src/index.ts
@@ -1,0 +1,1 @@
+export * from './job-proposal.plugin';
