+++ packages/plugins/job-proposal/src/proposal/dto/create-proposal.dto.ts
