+++ packages/plugins/job-proposal/src/proposal/dto/index.ts
