+++ packages/plugins/job-proposal/src/proposal/dto/proposal.dto.ts
