+++ packages/plugins/job-proposal/src/proposal/dto/update-proposal.dto.ts
