+++ packages/plugins/job-proposal/tsconfig.build.json
