+++ packages/plugins/job-proposal/tsconfig.json
