+++ packages/plugins/job-proposal/tsconfig.spec.json
