+++ packages/plugins/job-proposal/tslint.json
