--- packages/plugins/job-proposals/README.md
@@ -1,15 +1,0 @@
-# Job Proposals Plugin
-
-## Overview
-
-## Features
-
-## Installation
-
-To install the Job Proposals Plugin, simply run the following command in your terminal:
-
-```bash
-npm install @gauzy/job-proposals
-# or
-yarn add @gauzy/job-proposals
-```
