--- packages/plugins/job-proposals/src/index.ts
@@ -1,1 +1,0 @@
-export * from './job-proposals.plugin';
